# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi primera documentacion con swagger",
        "description": "Si",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Lista todos los productos.",
                "operationId": "453a21494655e2769474af4d0266186f",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crea un nuevo producto.",
                "operationId": "b93f6eb93990a81eb3c9d67e43d87083",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Precio",
                                    "Descripcion",
                                    "id_categoria",
                                    "id_marca"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Televisor"
                                    },
                                    "Precio": {
                                        "type": "integer",
                                        "example": 15000
                                    },
                                    "Descripcion": {
                                        "type": "string",
                                        "example": "Televisor de 50 pulgadas LG"
                                    },
                                    "id_categoria": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "id_marca": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtiene producto por ID.",
                "operationId": "02e54cf544a2dbd8d8bb9925df32b375",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualiza un  producto.",
                "operationId": "d9e4a80c4738d90111b188e69741233b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Precio",
                                    "Descripcion",
                                    "id_categoria",
                                    "id_marca"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Televisor"
                                    },
                                    "Precio": {
                                        "type": "integer",
                                        "example": 15000
                                    },
                                    "Descripcion": {
                                        "type": "string",
                                        "example": "Televisor de 50 pulgadas LG"
                                    },
                                    "id_categoria": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "id_marca": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Borra un producto.",
                "operationId": "a6375bfba84ef7579fa076d68d7250c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Producto": {
                "title": "Producto",
                "description": "Modelo de Producto",
                "properties": {
                    "id_producto": {
                        "description": "ID del producto",
                        "type": "integer",
                        "example": "1"
                    },
                    "nombre": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "example": "Camiseta"
                    },
                    "descripcion": {
                        "description": "Descripción del producto",
                        "type": "string",
                        "example": "Camiseta de algodón de manga corta"
                    },
                    "precio": {
                        "description": "Precio del producto",
                        "type": "number",
                        "format": "float",
                        "example": "29.99"
                    },
                    "id_categoria": {
                        "description": "ID de la categoría del producto",
                        "type": "integer",
                        "example": "1"
                    },
                    "id_marca": {
                        "description": "ID de la marca del producto",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        }
    }
}